requirements.txt
customtkinter
sounddevice
soundfile
numpy
faster-whisper
openai
python-dotenv
